# src/main/java/com/example/openapidemo/PetResource.java
package com.example.openapidemo;

public class PetResource {

    private long id;

    private String name;

    public long getId() {
        return id;
    }

    public void setId(long id) {
        this.id = id;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }
}